chore: small refactor and style tweaks
