+++ src/screens/settings/styles.ts
@@ -20,7 +20,7 @@ const Title = styled.Text`
 
 const SignOutButton = styled.TouchableOpacity`
   width: 100%;
-  background-color: ${colors.lowOpacity.black};
+  background-color: ${colors.lowOpacity.blackLow};
   margin-top: 18px;
   padding-vertical: 10px;
   border-radius: 2px;
