+++ src/theme/colors.ts
@@ -20,7 +20,8 @@ const colors = {
     whiteLow: 'rgba(255, 255, 255, 0.3)',
     whiteMid: 'rgba(255, 255, 255, 0.6)',
     blackSuperLow: 'rgba(255, 255, 255, 0.08)',
-    black: 'rgba(0, 0, 0, 0.2)',
+    blackLow: 'rgba(0, 0, 0, 0.2)',
+    blackMid: 'rgba(0, 0, 0, 0.6)',
     redMid: 'rgba(250,1,51, 0.6)',
     feedback: {
       positive: 'rgba(0, 255, 0, 0.2)',
